monitoring/requirements.txt
fastapi
uvicorn[standard]
pymongo
pika
python-socketio[asyncio_client]
python-dateutil
pydantic
python-dotenv
